requirements.txt
et-xmlfile==1.0.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
speedtest-cli==2.1.2
threaded==4.1.0
